refactor: folder restructuring v2
