+++ src/app/@modal/components/ModalSlot.tsx
@@ -4,7 +4,7 @@ import { useSelectedLayoutSegment } from "next/navigation";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 import { Dialog, DialogContent } from "../../../components/ui/dialog";
-import { useDialogActions } from "@/hooks/useDialog";
+import { useDialogActions } from "@/libs/hooks/useDialog";
 
 export default function ModalSlot({ children }: { children: React.ReactNode }) {
   const segment = useSelectedLayoutSegment("modal");
