+++ src/components/FormSubmissionLoader.tsx
@@ -1,6 +1,6 @@
 import { useFormStatus } from "react-dom";
 import { LoadingSpinner } from "./Spinner";
-import { cn } from "@/utils";
+import { cn } from "@/libs/utils";
 
 export function FormSubmissionLoader({ className }: { className?: string }) {
   // Note: useFormStatus works only in the context of a form
