+++ src/components/ui/button.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Slot } from "@radix-ui/react-slot";
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/utils/index";
+import { cn } from "@/libs/utils/index";
 
 // eslint-disable-next-line tailwindcss/no-custom-classname
 const buttonVariants = cva(
