+++ src/components/ui/dialog.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 import { X } from "lucide-react";
 
-import { cn } from "@/utils/index";
+import { cn } from "@/libs/utils/index";
 
 const Dialog = DialogPrimitive.Root;
