+++ src/components/ui/popover.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as PopoverPrimitive from "@radix-ui/react-popover";
 
-import { cn } from "@/utils/index";
+import { cn } from "@/libs/utils/index";
 
 const Popover = PopoverPrimitive.Root;
