+++ src/components/ui/tooltip.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 
-import { cn } from "@/utils/index";
+import { cn } from "@/libs/utils/index";
 
 const TooltipProvider = TooltipPrimitive.Provider;
